requirements.txt
alembic==1.4.2
appdirs==1.4.4
asn1crypto==1.3.0
audioread==2.1.8
Authlib==0.14.2
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.1
cryptography==2.9.2
Cython==0.29.17
decorator==4.4.2
distlib==0.3.0
elasticsearch==7.5.1
facebook-sdk==3.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-OAuth==0.12
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.25.0
gevent==20.5.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googletrans==2.4.0
greenlet==0.4.15
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
iteration-utilities==0.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jsonify==0.5
librosa==0.7.2
llvmlite==0.31.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.4.0
nose2==0.9.2
numba==0.48.0
numpy==1.18.4
oauth2==1.9.0.post1
oauthlib==3.1.0
pathlib==1.0.1
Pillow==7.1.2
pipenv==2018.11.26
protobuf==3.12.0
psycopg2-binary==2.8.5
pur==5.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
rarfile==3.1
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.23.0
scipy==1.4.1
sentry-sdk==0.14.4
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.17
stripe==2.48.0
threadpoolctl==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.4
zipp==3.1.0